splitted TimeMark
